MEN09-276 #close Maintenance [Fix Bugs]

+++ textFiles/clear.sql
@@ -171,8 +171,8 @@ CREATE TABLE pharmacy_product(
 
 CREATE TABLE delivery_run(
     delivery_run_id INTEGER GENERATED ALWAYS AS IDENTITY CONSTRAINT pk_delivery_run_id PRIMARY KEY,
-    distance float,
-    energy float,
+    distance number(8,3),
+    energy number(8,3),
     start_date timestamp,
     end_date timestamp NULL,
     courier_email varchar(255),
